[LCR-208] fix: type 분리해서 정의
