+++ src/types/MapScreenType.ts
@@ -1,14 +1,10 @@
-export type BannerItemType = {
-  image: number;
-  title: string;
-  subTitle: string;
-};
-
-export type PopUpItem = {
+export type popUpMarkerItem = {
   popupId: number;
   popupName: string;
   imageUrl: number; // 추후 string으로 변경
   popupOpenDate: string;
   popupCloseDate: string;
   address: string;
+  latitude: number;
+  longitude: number;
 };
